Supabase Buckets/requirements.txt
# Required packages for Supabase video upload
supabase>=2.0.0
python-dotenv>=1.0.0
